* moved splitDigi.cxx     * added sumHists.exe which sums histograms w/ same name accross multiple files
    * TrigHists class now derives from HistVec, trigger histograms are stored in subfolders now
    * added genTrigMonitor.exe which generates trigger threshold histograms from standard LAT configuration data

+++ calibGenCAL/src/CIDAC2ADC/IntNonlinAlg.cxx
@@ -269,7 +269,7 @@ namespace calibGenCAL {
         if (m_hugeTuple) {
           m_tupleData.rngIdx = rngIdx.val();
           m_tupleData.goodEvt = eventData.iGoodEvt;
-          m_tupleData.cidac = cidac;
+          m_tupleData.cidac = (unsigned short)cidac;
           m_tupleData.adc = adc;
           m_hugeTuple->Fill();
         }
